requirements.txt
anyio==3.6.1
click==8.1.3
fastapi==0.85.1
h11==0.14.0
idna==3.4
pydantic==1.10.2
sniffio==1.3.0
starlette==0.20.4
typing_extensions==4.4.0
uvicorn==0.18.3

pyJWT==2.5.0
python-decouple==3.6
pydantic[email]
SQLAlchemy==1.4.42
black==22.10.0